mysite/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
Brotli==1.1.0
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
duckduckgo_search==7.2.0
exceptiongroup==1.2.2
fake-useragent==2.0.3
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
gensim==4.3.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
lxml==5.3.0
MarkupSafe==3.0.2
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.2
node2vec==0.5.0
numpy==1.26.4
openai==1.59.3
primp==0.9.3
propcache==0.2.1
pydantic==2.10.4
pydantic_core==2.27.2
pymssql==2.3.2
pyodbc==5.2.0
python-dotenv==1.0.1
python-tds==1.16.0
requests==2.32.3
scikit-learn==1.6.0
scipy==1.13.1
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlalchemy_pytds==1.0.2
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.3
